refactor: move to the right folder the VIP 272
