--- simulations/vip-213/vip-213/abi/VTreasury.json
@@ -1,83 +1,0 @@
-[
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": true, "internalType": "address", "name": "previousOwner", "type": "address" },
-      { "indexed": true, "internalType": "address", "name": "newOwner", "type": "address" }
-    ],
-    "name": "OwnershipTransferred",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "address", "name": "tokenAddress", "type": "address" },
-      { "indexed": false, "internalType": "uint256", "name": "withdrawAmount", "type": "uint256" },
-      { "indexed": false, "internalType": "address", "name": "withdrawAddress", "type": "address" }
-    ],
-    "name": "WithdrawTreasuryBEP20",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "uint256", "name": "withdrawAmount", "type": "uint256" },
-      { "indexed": false, "internalType": "address", "name": "withdrawAddress", "type": "address" }
-    ],
-    "name": "WithdrawTreasuryBNB",
-    "type": "event"
-  },
-  { "payable": true, "stateMutability": "payable", "type": "fallback" },
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "owner",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [],
-    "name": "renounceOwnership",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [{ "internalType": "address", "name": "newOwner", "type": "address" }],
-    "name": "transferOwnership",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      { "internalType": "address", "name": "tokenAddress", "type": "address" },
-      { "internalType": "uint256", "name": "withdrawAmount", "type": "uint256" },
-      { "internalType": "address", "name": "withdrawAddress", "type": "address" }
-    ],
-    "name": "withdrawTreasuryBEP20",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      { "internalType": "uint256", "name": "withdrawAmount", "type": "uint256" },
-      { "internalType": "address payable", "name": "withdrawAddress", "type": "address" }
-    ],
-    "name": "withdrawTreasuryBNB",
-    "outputs": [],
-    "payable": true,
-    "stateMutability": "payable",
-    "type": "function"
-  }
-]
