+++ simulations/vip-272/bsctestnet.ts
@@ -2,9 +2,9 @@ import { expect } from "chai";
 import { BigNumber } from "ethers";
 import { ethers } from "hardhat";
 
-import { expectEvents } from "../../../src/utils";
-import { forking, testVip } from "../../../src/vip-framework";
-import { XVS, XVS_AMOUNT, XVS_BRIDGE_SRC, vip213Testnet } from "../../../vips/vip-213/vip-213-testnet";
+import { expectEvents } from "../../src/utils";
+import { forking, testVip } from "../../src/vip-framework";
+import { XVS, XVS_AMOUNT, XVS_BRIDGE_SRC, vip272 } from "../../vips/vip-272/bsctestnet";
 import COMPTROLLER_ABI from "./abi/Comptroller.json";
 import ERC20_ABI from "./abi/ERC20.json";
 import XVS_BRIDGE_ABI from "./abi/XVSProxyOFTSrc.json";
@@ -22,7 +22,7 @@ forking(36724865, () => {
     oldXVSBal = await xvs.balanceOf(XVS_BRIDGE_SRC);
   });
 
-  testVip("VIP-213 Send XVS to Dest Chain", vip213Testnet(), {
+  testVip("VIP-272 Send XVS to Dest Chain", vip272(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [COMPTROLLER_ABI, XVS_BRIDGE_ABI], ["VenusGranted", "SendToChain"], [1, 1]);
     },
